feat: Avatar 컴포넌트 구현 #114 - 기존 Avatar 컴포넌트를 GenderAvatar로 교체함
